% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate.corr.R
\name{bivariate.corr}
\alias{bivariate.corr}
\title{A Data Output Function}
\usage{
bivariate.corr(x, use = "pairwise", method = "pearson",
  adjust = "bonferroni")
}
\arguments{
\item{x}{dataframe}

\item{use}{"pairwise" or "complete"}

\item{method}{"pearson"}

\item{adjust}{"bonferroni"}
}
\description{
This function will perform and display a bi-variate correlation analysis
}
\examples{
bivariate.corr(x, use = "pairwise", method = "pearson", adjust = "bonferroni")
}
