% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_rows.R
\name{collapse_rows}
\alias{collapse_rows}
\title{Collapse rows}
\usage{
collapse_rows(df_g, col, lookleft = TRUE)
}
\arguments{
\item{x}{a table}
}
\description{
Taken from https://gist.github.com/jmclawson/6852c14023d7d9b7e91bbcfa419adeb8
}
