% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor.plot.R
\name{cor.plot}
\alias{cor.plot}
\title{A Results Output Function}
\usage{
cor.plot(x, use = "pairwise")
}
\arguments{
\item{x}{dataframe}

\item{use}{How to deal with missing data. Pairwise or listwise deletion (default = "pairwise")}
}
\description{
This function will display a Correlation Plot heat map
}
\examples{
cor.plot(x)
}
