% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.display.R
\name{data.display-deprecated}
\alias{data.display-deprecated}
\title{A Data Output Function}
\usage{
data.display(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display a scrollable data table
}
\seealso{
\code{\link{resultsoutput-deprecated}}
}
\keyword{internal}
