% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_display.R
\name{data_display}
\alias{data_display}
\title{A Data Output Function}
\usage{
data_display(x, round = 2, height = "800px", left.freeze = 1)
}
\arguments{
\item{x}{dataframe}

\item{round}{How many decimal places to round to?}

\item{height}{The height of the displayed table in px}

\item{left.freeze}{How many columns (starting from the left) to freeze}
}
\description{
This function will display a scrollable data table
}
\examples{
data_display(x)
}
