% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{descriptives}
\alias{descriptives}
\title{A Results Output Function}
\usage{
descriptives(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display basic descriptive statistics for a dataframe
}
\examples{
descriptives(x)
}
