% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_plot.R
\name{freq_plot}
\alias{freq_plot}
\title{Frequency distribution histograms for all variables}
\usage{
freq_plot(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will print out frequency distribution historgrams for each
variable in a dataframe
}
\examples{
freq_plot(data)
}
