% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\docType{data}
\name{GeomFlatViolin}
\alias{GeomFlatViolin}
\title{A Results Output Function}
\format{An object of class \code{GeomFlatViolin} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 6.}
\usage{
GeomFlatViolin
}
\arguments{
\item{mapping}{null}

\item{data}{null}

\item{stat}{ydensity}

\item{position}{dodge}

\item{trim}{true}

\item{scale}{aread}

\item{show.legend}{NA}

\item{inherit.aes}{TRUE}
}
\description{
This is a ggplot2 geom
}
\examples{
geom_flat_violin(x)
}
\keyword{datasets}
