% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coeff.R
\name{get_coeff}
\alias{get_coeff}
\title{Create Coefficients Data Frame}
\usage{
get_coeff(
  x,
  standardized = TRUE,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = NULL,
  effects = "all",
  digits = 3
)
}
\arguments{
\item{x}{an lm model object}

\item{standardized}{see regression_coeff()}

\item{ci}{see regression_coeff()}

\item{ci_method}{see regression_coeff()}

\item{bootstrap}{see regression_coeff()}

\item{iterations}{see regression_coeff()}

\item{effects}{"fixed" or "all" fixed and random effects.
default is "all"}

\item{digits}{see regression_coeff()}
}
\description{
Create Coefficients Data Frame
}
