% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{\%||\%}
\alias{\%||\%}
\title{A Results Output Function}
\usage{
a \%||\% b
}
\arguments{
\item{mapping}{null}

\item{data}{null}

\item{stat}{ydensity}

\item{position}{dodge}

\item{trim}{true}

\item{scale}{aread}

\item{show.legend}{NA}

\item{inherit.aes}{TRUE}
}
\description{
This is a ggplot2 geom
}
\examples{
geom_flat_violin(x)
}
