% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICC.print.R
\name{ICC.print}
\alias{ICC.print}
\title{A Results Output Function}
\usage{
ICC.print(x, random_effects = "Subject")
}
\arguments{
\item{x}{lmer object}

\item{random_effects}{random effects variable name (default = "Subject")}
}
\description{
This function will display the ICC value of an HLM model
}
\examples{
ICC.print(x)
}
