% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlm.output.R
\name{lmer.output}
\alias{lmer.output}
\title{A Results Output Function}
\usage{
lmer.output(x, random_effects = "Subject")
}
\arguments{
\item{x}{lmer object}

\item{random_effects}{random effects variable name (default = "Subject")}
}
\description{
This function will display the results of an lmer() HLM model
}
\examples{
lmer.output(x)
}
