% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.freq.R
\name{plot.freq-deprecated}
\alias{plot.freq-deprecated}
\title{A Data Output Function}
\usage{
plot.freq(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will print out frequency distribution historgrams for each variable in a dataframe
}
\seealso{
\code{\link{resultsoutput-deprecated}}
}
\keyword{internal}
