% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.freq.R
\name{plot.freq}
\alias{plot.freq}
\title{A Data Output Function}
\usage{
\method{plot}{freq}(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will print out frequency distribution historgrams for each variable in a dataframe
}
\examples{
plot.freq(data, x = "x.variable", y = "y.variable", se = TRUE, x.lim = c(NA,3), y.lim = c(0,100))
}
