% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.means.R
\name{plot.means}
\alias{plot.means}
\title{A Results Output Function}
\usage{
\method{plot}{means}(x, measurevar = "", withinvars = "",
  betweenvars = NULL, idvar = "", errorbars = "se",
  errorbars.color = "black", bar.color = "grey", x.label = "",
  y.label = "")
}
\arguments{
\item{x}{dataframe}

\item{measurevar}{a}

\item{withinvars}{a}

\item{betweenvars}{a}

\item{idvar}{a}

\item{errorbars}{a}

\item{errorbars.color}{a}

\item{bar.color}{a}

\item{y.label}{a}

\item{x.abel}{a}
}
\description{
This function will display a graph and table for between, within, or between-within mean group(s) comparison
}
\examples{
plot.means(x)
}
