% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.scatter.R
\name{plot.scatter}
\alias{plot.scatter}
\title{A Data Output Function}
\usage{
\method{plot}{scatter}(df, x = "", y = "", p.value = 0.05, se = FALSE,
  line.color = "dodgerblue4", point.color = "black", x.lim = "",
  y.lim = "", x.by = 1, y.by = 1, title = "", subtitle = "",
  caption = "", grp = "", point.show = TRUE)
}
\arguments{
\item{df}{dataframe}

\item{x}{x-axis variable}

\item{y}{y-axis variable}

\item{p.value}{probablity value to test for significance (default = .05)}

\item{se}{logical. Display standard error? (default = FALSE)}

\item{line.color}{default = "dodgerblue4"}

\item{point.color}{default = "black"}

\item{x.lim}{Lower and upper limits of x-axis (c(lower,upper)). If one of the limits should be defined automatically set to NA}

\item{y.lim}{Lower and upper limits of x-axis (c(lower,upper)). If one of the limits should be defined automatically set to NA}

\item{x.by}{Interval of tick marks for x-axis}

\item{y.by}{Interval of tick marks for y-axis}

\item{title}{Add a plot title}

\item{subtitle}{Add a plot subtitle}

\item{caption}{Add a plot caption}

\item{grp}{What is the grouping variable?}

\item{point.show}{Show data points?}
}
\description{
This function will print a scatter plot with the correlation value in the upper right
}
\examples{
plot.scatter(data, x = "x.variable", y = "y.variable", se = TRUE, x.lim = c(NA,3), y.lim = c(0,100))
}
