% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.timeseries.R
\name{plot.timeseries}
\alias{plot.timeseries}
\title{A Data Output Function}
\usage{
\method{plot}{timeseries}(df, x = "", y = "", se = "", x.lim = "",
  y.lim = "", x.by = 1, y.by = 1, title = "", subtitle = "",
  caption = "", grp = "", se.transparency = 0.1)
}
\arguments{
\item{df}{dataframe}

\item{x}{x-axis variable}

\item{y}{y-axis variable}

\item{se}{which variable to use for the standard error bars}

\item{x.lim}{Lower and upper limits of x-axis (c(lower,upper)). If one of the limits should be defined automatically set to NA}

\item{y.lim}{Lower and upper limits of x-axis (c(lower,upper)). If one of the limits should be defined automatically set to NA}

\item{x.by}{Interval of tick marks for x-axis}

\item{y.by}{Interval of tick marks for y-axis}

\item{title}{Add a plot title}

\item{subtitle}{Add a plot subtitle}

\item{caption}{Add a plot caption}

\item{grp}{What is the grouping variable?}

\item{se.transparency}{The level of transparance of error bars. 0 = completely transparent}
}
\description{
This function will print a scatter plot with the correlation value in the upper right
}
\examples{
plot.timeseries(data, x = "x.variable", y = "y.variable", se = TRUE, x.lim = c(NA,3), y.lim = c(0,100))
}
