% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_freq.R
\name{plot_freq}
\alias{plot_freq}
\title{A Data Output Function}
\usage{
plot_freq(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will print out frequency distribution historgrams for each variable in a dataframe
}
\examples{
plot_freq(data)
}
