% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_coeff.R
\name{regression_coeff}
\alias{regression_coeff}
\title{Regression Coefficients}
\usage{
regression_coeff(
  x,
  y = NULL,
  z = NULL,
  standardized = TRUE,
  unstandardized = TRUE,
  ci = 0.95,
  ci_method = NULL,
  bootstrap = FALSE,
  iterations = NULL,
  effects = "all",
  digits = 3
)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}

\item{standardized}{Logical, indicating whether or not to print standardized
estimates. Standardized estimates are based on "refit" of the model
on standardized data but it will not standardize categorical predictors.
Defualt is TRUE.}

\item{unstandardized}{Logical, indicating whether or not to print
unstandardized estimates. Default is TRUE.}

\item{ci}{Confidence Interval (CI) level. Default to 0.95 (95%)}

\item{ci_method}{Documention based on ?parameters::parameters.
Method for computing degrees of freedom for confidence
intervals (CI) and the related p-values. Allowed are following options
(which vary depending on the model class): "residual", "normal",
"likelihood", "satterthwaite", "kenward", "wald", "profile", "boot",
"uniroot", "ml1", "betwithin", "hdi", "quantile", "ci", "eti", "si",
"bci", or "bcai". See section Confidence intervals and approximation of
degrees of freedom in model_parameters() for further details.
When ci_method=NULL, in most cases "wald" is used then.}

\item{bootstrap}{Documention based on ?parameters::parameters.
Should estimates be based on bootstrapped model? If TRUE, then arguments
of Bayesian regressions apply (see also bootstrap_parameters()).}

\item{iterations}{Documention based on ?parameters::parameters.
The number of bootstrap replicates. This only apply in the case of
bootstrapped frequentist models.}

\item{effects}{"fixed" or "all" fixed and random effects.
default is "all"}

\item{digits}{How many decimal places to round to? Default is 3.}
}
\description{
Displays a table for regression coefficients.
     Multiple models can be added (x, y, and z).
}
