% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.pairs.R
\name{scatterplots.pairs-deprecated}
\alias{scatterplots.pairs-deprecated}
\title{A Data Output Function}
\usage{
scatterplots.pairs(data, x = c(), y = c(), all.pairs = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{x}{X-axis variables}

\item{y}{Y-axis variables}

\item{all.pairs}{Logical. If TRUE then will overwrite x, and y parameters
and just print out all combinations of paris in the dataframe. (Default: FALSE)}
}
\description{
This function will print out scatterplots of all pairs of variables in a dataframe.
Histograms on x and y variables are displayed on the margins, using ggMarginal()
Correlation statistics are printed at the bottom left of figure.
}
\seealso{
\code{\link{resultsoutput-deprecated}}
}
\keyword{internal}
