% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterplots.pairs.R
\name{scatterplots.pairs}
\alias{scatterplots.pairs}
\title{A Data Output Function}
\usage{
scatterplots.pairs(x, id.label = "")
}
\arguments{
\item{x}{dataframe}

\item{id.label}{column that contains subject IDs}
}
\description{
This function will print out scatterplots of all pairs of variables in a dataframe.
Histograms on x and y variables are displayed on the margins, using ggMarginal()
Correlation statistics are printed at the bottom left of figure.
}
\examples{
scatterplots.pairs(x, variable.list = colnames(x))
}
