% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.fitmeasures.R
\name{sem.fitmeasures}
\alias{sem.fitmeasures}
\title{A Results Output Function}
\usage{
sem.fitmeasures(x)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}
}
\description{
This function will display a table of Model fit measures
}
\examples{
sem.fitmeasures(x)
}
