% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.rsquared.R
\name{sem.rsquared}
\alias{sem.rsquared}
\title{A Results Output Function}
\usage{
sem.rsquared(x)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}
}
\description{
This function will display a table of R squared values (explained variance)
}
\examples{
sem.rsquared(x)
}
