% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sem.sig.R
\name{sem.sig}
\alias{sem.sig}
\title{A Results Output Function}
\usage{
sem.sig(x)
}
\arguments{
\item{x}{results from a cfa() or sem() lavaan model}
}
\description{
This function will display a table of Model significance tests
}
\examples{
sem.sig(x)
}
