% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.descriptives.R
\name{table.descriptives}
\alias{table.descriptives}
\title{A Results Output Function}
\usage{
table.descriptives(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display basic descriptive statistics for a dataframe
}
\examples{
table.descriptives(x)
}
