% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.means.R
\name{table.means}
\alias{table.means}
\title{A Results Output Function}
\usage{
table.means(x, measurevar, withinvars = NULL, betweenvars = NULL,
  idvar = NULL)
}
\arguments{
\item{x}{dataframe}

\item{measurevar}{a}

\item{withinvars}{a}

\item{betweenvars}{a}

\item{idvar}{a}

\item{errorbars}{a}

\item{errorbars.color}{a}

\item{bar.color}{a}

\item{x.abel}{a}

\item{y.label}{a}
}
\description{
This function will display a table and table for between, within, or between-within mean group(s) comparison
}
\examples{
table.means(x)
}
