% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_coeff.R
\name{table_coeff}
\alias{table_coeff}
\title{Regression Coefficients}
\usage{
table_coeff(x, y = NULL, z = NULL)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}
}
\description{
Displays a table for regression coefficients using sjPlot::tab_model().
Multiple models can be added (x, y, and z).
}
