% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_corr.test.R
\name{table_corr.test}
\alias{table_corr.test}
\title{A Results Output Function}
\usage{
table_corr.test(x)
}
\arguments{
\item{x}{an object from psych::corr.test()}
}
\description{
This function will display basic descriptive statistics for a dataframe
}
\examples{
table_corr.test(x)
}
