% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_corr.test.R, R/table_descriptives.R
\name{table_descriptives}
\alias{table_descriptives}
\title{A Results Output Function}
\usage{
table_descriptives(x)

table_descriptives(x)
}
\arguments{
\item{x}{an object from psych::corr.test()}

\item{x}{dataframe}
}
\description{
This function will display basic descriptive statistics for a dataframe

This function will display basic descriptive statistics for a dataframe
}
\examples{
table_descriptives(x)
table_descriptives(x)
}
