% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_fit.R
\name{table_fit}
\alias{table_fit}
\title{Regression Model Fit}
\usage{
table_fit(x, y = NULL, z = NULL, print = TRUE)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}

\item{print}{Create a knitr table for displaying as html table?
(default = TRUE)}
}
\description{
Displays a table for the fit of regression models. Multiple models can
be added (x, y, and z).
}
