% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_model.R
\name{table_model}
\alias{table_model}
\title{Regression Model Tables}
\usage{
table_model(x, y = NULL, z = NULL, print = TRUE)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}

\item{print}{Create a knitr table for displaying as html table?
(default = TRUE)}
}
\description{
Display all tables relevant for Regression models. You can include up to
three models (x, y, and z) for hierarchical regression results.
}
