% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_rsquared.R
\name{table_rsquared}
\alias{table_rsquared}
\title{R-Squared Values}
\usage{
table_rsquared(x, y = NULL, z = NULL, print = TRUE)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}

\item{print}{Create a knitr table for displaying as html table?
(default = TRUE)}
}
\description{
Display a table of R-Squared values. If more than one model is added
(x, y, and z), then R-Squared and relevant stats will be displayed.
}
