% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_spacious.R
\name{theme_spacious}
\alias{theme_spacious}
\title{Spacious ggplot2 theme}
\usage{
theme_spacious(font.size = 14, bold = TRUE)
}
\arguments{
\item{font.size}{font size}

\item{bold}{bold labels?}
}
\description{
This will increase the distance between axis titles and axis labels
}
