% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_comparisons.R
\name{anova_comparisons}
\alias{anova_comparisons}
\title{ANOVA F-Table for Overall Main Effect and Interaction Terms}
\usage{
anova_comparisons(x, term = NULL, digits = 3)
}
\arguments{
\item{x}{an lmer model object}

\item{term}{The factor at which to compare levels at}

\item{digits}{How many decimal places to round to? Default is 3.}
}
\description{
ANOVA F-Table for Overall Main Effect and Interaction Terms
}
