% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_comparisons.R
\name{anova_comparisons}
\alias{anova_comparisons}
\title{Post-hoc Comparison Tables}
\usage{
anova_comparisons(
  x,
  contrast = NULL,
  at = NULL,
  adjust = "none",
  digits = 3,
  pbkrtest.limit = NULL,
  lmerTest.limit = NULL
)
}
\arguments{
\item{x}{an lmer model object}

\item{contrast}{The factor at which to compare levels at}

\item{at}{A second interacting factor to compare the effect of contrast at}

\item{adjust}{The p-values adjustment method for frequentist multiple
comparisons. Can be one of "holm", "tukey", "hochberg",
"hommel", "bonferroni", "BH", "BY", "fdr" or "none" (default). See the
p-value adjustment section in the emmeans::test documentation.}

\item{digits}{How many decimal places to round to? Default is 3.}

\item{pbkrtest.limit}{Optional parameter that can be set to help calculate dfs.
If you need to use this a warning message will appear in the console
telling you what to set this at.}

\item{lmerTest.limit}{Optional parameter that can be set to help calculate dfs.
If you need to use this a warning message will appear in the console
telling you what to set this at.}
}
\description{
Post-hoc Comparison Tables
}
