% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anova_modelsig.R
\name{anova_modelsig}
\alias{anova_modelsig}
\title{ANOVA F-Table for Overall Main Effect and Interaction Terms}
\usage{
anova_modelsig(x, digits = 3, id_col = "Subject")
}
\arguments{
\item{x}{an lmer model object}

\item{digits}{How many decimal places to round to? Default is 3.}

\item{id_col}{The column containing subject ids. Default is "Subject"}
}
\description{
ANOVA F-Table for Overall Main Effect and Interaction Terms
}
