% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{descriptives_table}
\alias{descriptives_table}
\title{Descriptive statistics table}
\usage{
descriptives_table(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display basic descriptive statistics for a dataframe
}
