% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_table.R
\name{descriptives_table}
\alias{descriptives_table}
\title{Descriptive Statistics Table}
\usage{
descriptives_table(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
This function will display a table of descriptive statistics
}
